requirements.txt
pandas
streamlit
requests
sklearn.feature_extraction.text
sklearn.metrics.pairwise